EOS-16520 : Convert sspl_setup shell script to python and Renamed bin source directory to setup directory
