+++ jenkins/build.sh
@@ -22,7 +22,7 @@ BASE_DIR=$(realpath $(dirname $0)/..)
 PARENT_DIR=$(basename "$BASE_DIR")
 
 PROG_NAME=$(basename $0)
-source "$BASE_DIR/low-level/files/opt/seagate/sspl/bin/constants.sh"
+source "$BASE_DIR/low-level/files/opt/seagate/sspl/setup/constants.sh"
 
 RPM_BUILD_PATH=${DIST:-$HOME/rpmbuild}
