+++ low-level/files/opt/seagate/sspl/setup/conf_based_sensors_enable
