+++ low-level/files/opt/seagate/sspl/setup/generate_resource_health_view/resource_health_view
